Service region almost done

+++ services/src/main/java/com/carswaddle/services/services/serviceModels/TaxInfo.kt
@@ -2,6 +2,6 @@ package com.carswaddle.carswaddleandroid.services.serviceModels
 
 data class TaxInfo (
     val taxYear: String,
-    val metersDriven: Int,
-    val mechanicCostInCents: Int,
+    val metersDriven: Double,
+    val mechanicCostInCents: Double,
 )
